requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright>=1.40.0
pandas>=2.1.0
openpyxl>=3.1.0
reportlab>=4.0.0
python-dotenv>=1.0.0
